Move Datadog.Trace.TestHelpers.Result to its own file and move the constructor to the correct order after removing extra compiler directives
